+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### 10th September 2024 - Version 1.2.9
+
+-   Tweak: Display user error message when the event map code is incorrect
+
 #### 22th April 2024 - Version 1.2.8
 
 -   Fix: Error message in debug mode when artist taxonomy does not exist